django--recipes/pyproject.toml
[tool.poetry]
name = "recipebook"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.6"
django-extensions = "^2.2.9"
django-debug-toolbar = "^2.2"
werkzeug = "^1.0.1"
django-environ = "^0.4.5"
django-registration-redux = "^2.8"
django-ordered-model = "^3.4.1"

[tool.poetry.dev-dependencies]
prospector = "^1.2.0"
mypy = "^0.770"
yapf = "^0.30.0"
pytest = "^5.4.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
